sensor/gesture/ollama: fix 13 issues from code review
Correctness:
- contact: add TWIST_TORQUE_MAX guard to hold-path twist entry/exit (matched no-hold path)
- stroke: gate HoldDetector on MIN_WINDOW_FRAMES to prevent early hold misclassification
- stroke: add R² floor (0.25) to StrokeDetector to reject noisy non-linear centroid paths
- ollama: add is_active() delegation so autonomous patterns suppress gesture classifier
- ollama: handle promoted events by updating matching started event in batch (not silently dropped)

Robustness:
- ollama: wrap _apply_llm_response in try/except; move clear_history to finally
- ollama: defensive float() conversion for LLM speed field with warning log
- client: fix unreachable except TimeoutError — detect via isinstance(exc.reason, TimeoutError)

Prompt / gesture encoding:
- ollama: add intensity label (light/firm) to all gesture types via _categorize_gesture
- ollama: handle cradle gesture type with duration qualifier
- robot_context.md: update gesture reference with intensity suffix and cradle definitions

Cleanup:
- client: remove dead _prune() and _turn_times (clear_history resets after every call)
- ollama: move _update_pattern_params inside lock for correctness
- ollama: fix docstring claiming persistent history (it clears after every call)
- patterns: fix CREST_POWER type annotation int→float

+++ petctl/schemes/patterns.py
@@ -1842,7 +1842,7 @@ class PurrRippleMotion(Motion):
 
     KD_TARGET: float = 0.08  # peak kd — causes vibration
     BASE_HZ: float = 0.3    # ripple frequency at speed=1.0 (~3.3s per pass)
-    CREST_POWER: int = 0.5    # exponent on sin envelope; higher = narrower active crest
+    CREST_POWER: float = 0.5    # exponent on sin envelope; fractional values narrow the crest
 
     def __init__(self, speed: float = 1.0) -> None:
         self.speed = max(speed, 0.01)
